Github ReRun Command
Implementation of support for re-run the check suite.
We created a base code to be shared between ReRun
through comments and the command sent by github.
